Move databrowser app generic-renderer into its own module
