+++ databrowser/modules/generic-renderer/components/GenericDataRenderer.vue
