+++ databrowser/modules/generic-renderer/components/GenericFilterRenderer.vue
