+++ databrowser/modules/generic-renderer/components/generic-renderer.css
