+++ databrowser/modules/generic-renderer/config/generic-render.config.ts
