+++ databrowser/modules/generic-renderer/config/renderer.enum.ts
