+++ databrowser/nuxt.config.js
@@ -21,8 +21,12 @@ export default {
 
   // Plugins to run before rendering page: https://go.nuxtjs.dev/config-plugins
   plugins: [
-    { src: '~/plugins/renderer.web-component.import.ts' },
-    { src: '~/plugins/generic.web-component.import.ts' },
+    {
+      src: '~/modules/generic-renderer/plugins/renderer.web-component.import.ts',
+    },
+    {
+      src: '~/modules/generic-renderer/plugins/generic.web-component.import.ts',
+    },
   ],
 
   // Auto import components: https://go.nuxtjs.dev/config-components
